refactor: Refactor Board Controller and Services for Improved Access Control
- Introduced BoardAccessVerifier to streamline board access control checks.
- Updated getBoardByUserAndId in BoardService to simplify visibility and ownership checks.
- Enhanced GlobalExceptionHandler to handle UserNotAuthenticatedException alongside AuthenticationException.
- Removed redundant visibility checks in service methods to centralize access control logic.

These changes improve code readability and maintainability while ensuring consistent access control across board-related operations.

+++ backend/src/main/java/sit/int221/backend/controllers/BoardStatusController.java
@@ -9,6 +9,7 @@ import sit.int221.backend.dtos.AddEditStatusDTO;
 import sit.int221.backend.dtos.StatusDTO;
 import sit.int221.backend.project_management.Status;
 import sit.int221.backend.services.BoardStatusService;
+import sit.int221.backend.utils.BoardAccessVerifier;
 
 import java.util.List;
 
@@ -17,34 +18,41 @@ import java.util.List;
 @AllArgsConstructor
 public class BoardStatusController {
     private final BoardStatusService boardStatusService;
+    private BoardAccessVerifier boardAccessVerifier;
 
     @GetMapping
     public ResponseEntity<List<Status>> findAllByBoardId(@PathVariable String boardId) {
+        boardAccessVerifier.verifyUserBoardAccess(boardId, false);
         return ResponseEntity.ok(boardStatusService.findAllByBoardId(boardId));
     }
 
     @GetMapping("/{statusId}")
     public ResponseEntity<Status> findByIdAndBoardId(@PathVariable Integer statusId, @PathVariable String boardId) {
+        boardAccessVerifier.verifyUserBoardAccess(boardId, false);
         return ResponseEntity.ok(boardStatusService.findByIdAndBoardId(statusId, boardId));
     }
 
     @PostMapping
     public ResponseEntity<Status> createStatus(@Valid @RequestBody AddEditStatusDTO statusDTO, @PathVariable String boardId) {
+        boardAccessVerifier.verifyUserBoardAccess(boardId, true);
         return ResponseEntity.status(HttpStatus.CREATED).body(boardStatusService.createStatus(statusDTO, boardId));
     }
 
     @PutMapping("/{statusId}")
     public ResponseEntity<StatusDTO> updateStatusById(@PathVariable Integer statusId, @PathVariable String boardId, @Valid @RequestBody AddEditStatusDTO statusDTO) {
+        boardAccessVerifier.verifyUserBoardAccess(boardId, true);
         return ResponseEntity.ok(boardStatusService.updateStatus(statusId, boardId, statusDTO));
     }
 
     @DeleteMapping("/{statusId}")
     public ResponseEntity<Status> deleteStatusById(@PathVariable Integer statusId, @PathVariable String boardId) {
+        boardAccessVerifier.verifyUserBoardAccess(boardId, true);
         return ResponseEntity.ok(boardStatusService.deleteStatusById(statusId, boardId));
     }
 
     @DeleteMapping("/{statusId}/{newStatusId}")
     public ResponseEntity<Status> transferAndDeleteStatus(@PathVariable Integer statusId, @PathVariable Integer newStatusId, @PathVariable String boardId) {
+        boardAccessVerifier.verifyUserBoardAccess(boardId, true);
         Status transferAndDeleteStatus = boardStatusService.transferAndDeleteStatus(statusId, newStatusId, boardId);
         return ResponseEntity.ok(transferAndDeleteStatus);
     }
